requirements.txt
requests==2.23.0
matplotlib==3.3.2
tensorflow_gpu==2.3.1
Keras==2.4.3
numpy==1.18.5
Pillow==8.0.1
tensorflow==2.3.1
fastapi==0.62.0 
uvicorn==0.13.0
